# swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/investment": {
      "get": {
        "operationId": "InvestmentController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateInvestmentDtoResult"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "investment"
        ]
      },
      "post": {
        "operationId": "InvestmentController_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "investment"
        ]
      }
    },
    "/investment/{id}": {
      "delete": {
        "operationId": "InvestmentController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "investment"
        ]
      }
    },
...
  },
  "info": {
    "title": "Investment Advisor",
    "description": "Assists in choosing fixed-income investments: CDB, RDB, LCI, LCA, and Savings Account.",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [
    {
      "url": "https://li-invest.koyeb.app"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateInvestmentDtoResult": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Tipo da aplicação",
            "examples": [
              "LCA",
              "LCI",
              "CDB"
            ]
          },
          "issuer": {
            "type": "string",
            "description": "Emissor",
            "example": "Banco Itaú"
          },
          "originBank": {
            "type": "string",
            "description": "Banco que disponibiliza o investimento",
            "example": "Itaú"
          },
          "minimumApplication": {
            "type": "number",
            "description": "Aplicação Mínima",
            "example": 100
          },
          "maturity": {
            "type": "string",
            "description": "Vencimento",
            "example": "2025-12-31",
            "format": "date"
          },
          "profitability": {
            "type": "number",
            "description": "Rentabilidade",
            "example": 100
          },
          "profitabilityType": {
            "type": "string",
            "description": "Rentabilidade do",
            "example": "CDI",
            "default": "CDI"
          },
          "id": {
            "type": "number",
            "description": "ID"
          }
        },
        "required": [
          "originBank",
          "minimumApplication",
          "maturity",
          "profitability",
          "profitabilityType",
          "id"
        ]
      },
      "CreateInvestmentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Tipo da aplicação",
            "examples": [
              "LCA",
              "LCI",
              "CDB"
            ]
          },
          "issuer": {
            "type": "string",
            "description": "Emissor",
            "example": "Banco Itaú"
          },
          "originBank": {
            "type": "string",
            "description": "Banco que disponibiliza o investimento",
            "example": "Itaú"
          },
          "minimumApplication": {
            "type": "number",
            "description": "Aplicação Mínima",
            "example": 100
          },
          "maturity": {
            "type": "string",
            "description": "Vencimento",
            "example": "2025-12-31",
            "format": "date"
          },
          "profitability": {
            "type": "number",
            "description": "Rentabilidade",
            "example": 100
          },
          "profitabilityType": {
            "type": "string",
            "description": "Rentabilidade do",
            "example": "CDI",
            "default": "CDI"
          }
        },
        "required": [
          "originBank",
          "minimumApplication",
          "maturity",
          "profitability",
          "profitabilityType"
        ]
      }
    }
  }
}